+++ src/EnlighterJS.Gutenberg.js
@@ -14,7 +14,7 @@ import _inlineFormats from './blocks/richtext/formats';
 Blocks.registerBlockType('enlighter/codeblock', _codeblock);
 
 // Register Enlighter formats
-//RichText.registerFormatType('enlighter/richtext-inline', _inlineFormats);
+RichText.registerFormatType('enlighter/richtext-inline', _inlineFormats);
 
 // static properties
 export const version = '[[VERSION]]';